+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `.codesandbox` tasks and `.devcontainer` config files.
 - Added At icons.
 - Add Battery icons.
+- Add Book Closed icons.
+- Add Book Open icons.
 
 ## [0.1.3] - 2023-11-04